requirements.txt
python-dotenv
rich
InquirerPy
babel
amadeus